mm: zswap: Set shrinker batch size to 1024
Modify the zswap shrinker to use a batch size of 1024 instead of 0. This
change allows the shrinker to process objects in larger batches, potentially
improving performance by reducing the overhead of frequent shrinker
invocations under memory pressure. The value 1024 is chosen to balance
throughput and responsiveness, aligning with common batch sizes used in
similar kernel subsystems.

Signed-off-by: brokestar233 <3765589194@qq.com>